💉 Set up Hilt DI based around AddDomainNameRuleViewModel
