+++ app/src/main/java/com/suvanl/fixmylinks/FmlApplication.kt
@@ -3,7 +3,9 @@ package com.suvanl.fixmylinks
 import android.app.Application
 import com.suvanl.fixmylinks.di.AppContainer
 import com.suvanl.fixmylinks.di.AppDataContainer
+import dagger.hilt.android.HiltAndroidApp
 
+@HiltAndroidApp
 class FmlApplication : Application() {
 
     lateinit var container: AppContainer
