+++ app/src/main/java/com/suvanl/fixmylinks/di/AppDataContainer.kt
@@ -2,10 +2,11 @@ package com.suvanl.fixmylinks.di
 
 import android.content.Context
 import com.suvanl.fixmylinks.data.local.db.RuleDatabase
+import com.suvanl.fixmylinks.data.repository.CustomRulesRepository
 import com.suvanl.fixmylinks.data.repository.RulesRepository
 
 class AppDataContainer(private val context: Context) : AppContainer {
     override val rulesRepository: RulesRepository by lazy {
-        RulesRepository(RuleDatabase.getDatabase(context = context))
+        CustomRulesRepository(RuleDatabase.getDatabase(context = context))
     }
 }
