+++ app/src/main/java/com/suvanl/fixmylinks/ui/util/ViewModelComposableUtil.kt
@@ -1,6 +1,7 @@
 package com.suvanl.fixmylinks.ui.util
 
 import androidx.compose.runtime.Composable
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.suvanl.fixmylinks.di.AppViewModelProvider
 import com.suvanl.fixmylinks.domain.mutation.MutationType
@@ -16,9 +17,7 @@ import com.suvanl.fixmylinks.viewmodel.newruleflow.AddRuleViewModel
 fun getNewRuleFlowViewModel(mutationType: MutationType): AddRuleViewModel {
     return when (mutationType) {
         MutationType.DOMAIN_NAME -> {
-            viewModel<AddDomainNameRuleViewModel>(
-                factory = AppViewModelProvider.Factory
-            )
+            hiltViewModel<AddDomainNameRuleViewModel>()
         }
 
         MutationType.URL_PARAMS_ALL -> {
