+++ build.gradle.kts
@@ -3,4 +3,12 @@ plugins {
     id("com.android.application") version "8.1.2" apply false
     id("org.jetbrains.kotlin.android") version "1.9.20" apply false
     id("com.guardsquare.appsweep") version "latest.release" apply false
+    id("com.google.dagger.hilt.android") version "2.44" apply false
+    id("com.google.devtools.ksp") version "1.9.20-1.0.14" apply false
+}
+
+buildscript {
+    dependencies {
+        classpath("com.google.dagger:hilt-android-gradle-plugin:2.48.1")
+    }
 }
